refactor(#3038): extract SensitivityState to voice_sensitivity sibling
Move the live barge-in sensitivity sub-struct (`SensitivityState`:
definition + impl) out of the `voice_barge_in.rs` giant into a new
`src/services/discord/voice_sensitivity.rs` sibling, continuing the
god-object split (after AcknowledgementConfig #3237, ConfigSnapshotCache
#3242, VoiceIdSequences #3244).

Why this sub-struct: highest cohesion / shallowest external coupling of
the remaining in-file candidates. Its three fields and six methods fully
encapsulate the concern; all 8 call-sites go through methods (no leaked
field access), and it depends only on already-public external types
(`BargeInSensitivity` / `BargeInSensitivityState`). The other candidates
are more entangled: `StreamingSttSessions` is keyed by the in-file
private `StreamingSttKey` and exposes `&DashMap` (entry guards held
across await), and `VoiceProgressChannelState` is woven into the
progress-driver methods that remain in-file.

Concurrency preserved byte-for-byte: the `AtomicU8` mirror is stored
*before* the `RwLock` write (F18, #2046), `current()` falls back to the
atomic mirror (not the boot default) on `try_read` contention, the exact
`Relaxed` memory Ordering on store/load is unchanged, and
`state_handle()` clones the same `Arc<RwLock<_>>` for the TTL-reset task.
`VoiceBargeInRuntime` keeps the `sensitivity: SensitivityState` field and
only calls into it. Visibility minimized to
`pub(in crate::services::discord)`.

LoC: voice_barge_in.rs production 4728 -> 4657 (-71). mod.rs +1 for the
module registration. Baseline + change-surfaces freeze synced.

Gates: cargo fmt --check clean; cargo check --lib / --lib --tests clean
(0 new warnings); cargo test --lib voice_barge_in 52 passed; giant
ratchet + agent-maintenance freshness passed; ci-script-checks exit 0.

Refs #3038
