+++ src/services/discord/mod.rs
@@ -95,6 +95,7 @@ mod voice_barge_in;
 mod voice_config_cache;
 mod voice_id_sequences;
 mod voice_routing;
+mod voice_sensitivity;
 #[path = "watchers/lifecycle_decision.rs"]
 mod watcher_lifecycle_decision;
